bug fix dashboard

+++ tabTeam.php
@@ -3,8 +3,8 @@
 	include "file.php";
     require_once 'Database.php';
 	// use app\Database;
-    $objdb=new Database('psybotechnologies.com','psyboysg_test','psybotest','psyboysg_psybo-db');
-    // $objdb= new Database ('localhost','root','asd','psybo-db');
+    // $objdb=new Database('psybotechnologies.com','psyboysg_test','psybotest','psyboysg_psybo-db');
+    $objdb= new Database ('localhost','root','asd','psybo-db');
     $objfile=new File();
     $emp_id=$objdb->num_row_emp();// number of values of employee
     $count_emp=count($emp_id);
@@ -40,8 +40,10 @@
 		{
 		 	unlink(getcwd().$actdir.$file_name);
 		}
-		$objdb->delete_team($_GET['deleteid']);
-		// header("location:tabTeam.php");
+		if (!empty($ptf_id)) 
+		{
+			$objdb->delete_team($_GET['deleteid']);
+		}
 		if ($objdb == true ) 
 		{
 			$message ="<script type='text/javascript'>//alert('Deleted!');
